change startValue parameter of InverseCdf::set_defaults from int to double
This is to avoid data loss during an unnecessary type conversion.

+++ src/InverseCdf.hpp
@@ -8,7 +8,7 @@ class InverseCdf
 
 public:
 
-	void set_defaults(int startValue);
+	void set_defaults(double startValue);
 
 	void apply_exponent();
 	void apply_exponent(double value);
